Move TaskExtensions -> Microosft.VisualStudio.Threading.Tasks
